Informed flow agents

+++ abides-markets/abides_markets/utils/__init__.py
@@ -1,15 +1,13 @@
 import datetime
-import sys
-import traceback
 import warnings
 from contextlib import contextmanager
 from typing import List, Union
 
 import numpy as np
 import pandas as pd
+from abides_core import LatencyModel
 from scipy.spatial.distance import pdist, squareform
 
-from abides_core import LatencyModel
 
 # Utility method to flatten nested lists.
 def delist(list_of_lists):
@@ -72,7 +70,7 @@ def ignored(warning_str, *exceptions):
 
 
 def generate_uniform_random_pairwise_dist_on_line(
-    left: float, right: float, num_points: int, random_state: np.random.RandomState
+        left: float, right: float, num_points: int, random_state: np.random.RandomState
 ) -> np.ndarray:
     """Uniformly generate points on an interval, and return numpy array of pairwise
     distances between points.
@@ -162,7 +160,7 @@ def generate_latency_model(agent_count, latency_type="deterministic"):
         "no_latency",
     ], "Please select a correct latency_type"
 
-    latency_rstate = np.random.RandomState(seed=np.random.randint(low=0, high=2 ** 32))
+    latency_rstate = np.random.RandomState(seed=np.random.randint(low=0, high=2 ** 32, dtype=np.int64))
     pairwise = (agent_count, agent_count)
 
     if latency_type == "deterministic":
